pyproject.toml
[project]
name = "adithyasean"
version = "0.1.0"
description = "Drop-in ollama client with Cloudflare Access authentication via .env"
requires-python = ">=3.11"
dependencies = [
    "ollama>=0.4",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
